feat: rename page Sign Up to Early Access
