+++ src/components/pages/early-access/hero/images/illustration.png
