+++ src/components/pages/home/hero/hero.jsx
@@ -89,7 +89,7 @@ const Hero = () => {
           <Button
             id="hero-button"
             className="mt-10 2xl:mt-8 xl:mt-7 md:mt-6"
-            to={LINKS.signUp}
+            to={LINKS.earlyAccess}
             size="md"
             theme="primary"
           >
