+++ src/components/pages/home/saas/saas.jsx
@@ -44,7 +44,7 @@ const SaaS = () => {
           <Button
             id="saas-button"
             className="mt-10 2xl:mt-8 xl:mt-7 md:mt-6"
-            to={LINKS.signUp}
+            to={LINKS.earlyAccess}
             size="md"
             theme="primary"
           >
